Fatras: move Interactor into the detail namespace
